CDRIVER-520 On Apple, use native SecureTransport over OpenSSL
