+++ src/mongoc/mongoc-rand-openssl-private.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 MongoDB, Inc.
+ * Copyright 2015 MongoDB, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,27 +14,32 @@
  * limitations under the License.
  */
 
+#ifndef MONGOC_RAND_OPENSSL_PRIVATE_H
+#define MONGOC_RAND_OPENSSL_PRIVATE_H
 
 #if !defined (MONGOC_I_AM_A_DRIVER) && !defined (MONGOC_COMPILATION)
 #error "Only <mongoc.h> can be included directly."
 #endif
 
-
-#ifndef MONGOC_RAND_PRIVATE_H
-#define MONGOC_RAND_PRIVATE_H
+#ifdef MONGOC_OPENSSL
 
 #include <bson.h>
 
-#include "mongoc-rand.h"
-#include "mongoc-rand-apple-private.h"
-#include "mongoc-rand-openssl-private.h"
-
 BSON_BEGIN_DECLS
 
-int _mongoc_rand_bytes(uint8_t * buf, int num);
-int _mongoc_pseudo_rand_bytes(uint8_t * buf, int num);
+int
+_mongoc_rand_openssl_bytes (uint8_t *buf,
+                            int      num);
+
+int
+_mongoc_pseudo_rand_openssl_bytes (uint8_t *buf,
+                                   int      num);
 
 BSON_END_DECLS
 
+/* API setup for OpenSSL */
+#define _mongoc_rand_bytes_impl _mongoc_rand_openssl_bytes
+#define _mongoc_pseudo_rand_bytes_impl _mongoc_pseudo_rand_openssl_bytes
 
-#endif /* MONGOC_RAND_PRIVATE_H */
+#endif /* MONGOC_OPENSSL */
+#endif /* MONGOC_RAND_OPENSSL_PRIVATE_H */
