+++ src/mongoc/mongoc-ssl-apple-private.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 MongoDB, Inc.
+ * Copyright 2013 MongoDB, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,27 +14,34 @@
  * limitations under the License.
  */
 
+#ifndef MONGOC_SSL_APPLE_PRIVATE_H
+#define MONGOC_SSL_APPLE_PRIVATE_H
+
+#ifdef MONGOC_APPLE_NATIVE_TLS
 
 #if !defined (MONGOC_I_AM_A_DRIVER) && !defined (MONGOC_COMPILATION)
 #error "Only <mongoc.h> can be included directly."
 #endif
 
-
-#ifndef MONGOC_RAND_PRIVATE_H
-#define MONGOC_RAND_PRIVATE_H
-
-#include <bson.h>
-
-#include "mongoc-rand.h"
-#include "mongoc-rand-apple-private.h"
-#include "mongoc-rand-openssl-private.h"
+#include "mongoc-ssl.h"
 
 BSON_BEGIN_DECLS
 
-int _mongoc_rand_bytes(uint8_t * buf, int num);
-int _mongoc_pseudo_rand_bytes(uint8_t * buf, int num);
+char    *
+_mongoc_ssl_apple_extract_subject (const char *filename);
+
+void
+_mongoc_ssl_apple_init (void);
+
+void
+_mongoc_ssl_apple_cleanup (void);
+
+/* API setup for SecureTransport */
+#define _mongoc_ssl_extract_subject_impl _mongoc_ssl_apple_extract_subject
+#define _mongoc_ssl_init_impl _mongoc_ssl_apple_init
+#define _mongoc_ssl_cleanup_impl _mongoc_ssl_apple_cleanup
 
 BSON_END_DECLS
 
-
-#endif /* MONGOC_RAND_PRIVATE_H */
+#endif /* MONGOC_APPLE_NATIVE_TLS */
+#endif /* MONGOC_SSL_APPLE_PRIVATE_H */
