+++ src/mongoc/mongoc-stream-private.h
@@ -27,13 +27,13 @@
 
 BSON_BEGIN_DECLS
 
-
-#define MONGOC_STREAM_SOCKET   1
-#define MONGOC_STREAM_FILE     2
-#define MONGOC_STREAM_BUFFERED 3
-#define MONGOC_STREAM_GRIDFS   4
-#define MONGOC_STREAM_TLS      5
-
+#define MONGOC_STREAM_SOCKET      1
+#define MONGOC_STREAM_FILE        2
+#define MONGOC_STREAM_BUFFERED    3
+#define MONGOC_STREAM_GRIDFS      4
+#define MONGOC_STREAM_TLS         5
+#define MONGOC_STREAM_OPENSSL_TLS 6
+#define MONGOC_STREAM_APPLE_TLS   7
 
 BSON_END_DECLS
